Lots to do
I confused the terms "Plate" and "Menu." Tasks should start in the menu,
and move to the plate, but I did a little bit of both. They do share a
lot of code, but I did that because I wanted them to be static, so I
wouldn't have to keep track of instances. The pickling stuff may not be
right. Alternatively, we could write files with keywords and some sort
of start and stop key, and loop through the file, from start to finish,
for every keyword we want, find the keyword, and read everything between
the start and stop keys as the corresponding variable. As you can see,
it needs some work.